==> 1a_Scrublet.ipynb <==
# script for the implementation of scrublet to predict and remove doublets from single cell data
# By Louise Baldwin

# import packages
import scrublet as scr
import scipy.io
import matplotlib.pyplot as plt
import numpy as np
import os
import scanpy as sc

# directories
os.chdir("/share/ScratchGeneral/loubal/projects/MSC/mouse-single-cell")
in_file = ("data/processed/updated_merged.h5ad")
results_file = ("data/processed/merged_scrublet.h5ad")
figdir = ("outs/1a_Scrublet/figures/")
tabdir = ("outs/1a_Scrublet/tables/")
os.makedirs(figdir, exist_ok=True)
os.makedirs(tabdir, exist_ok=True)

adata=sc.read_h5ad(in_file)

adata

sc.external.pp.scrublet(adata, expected_doublet_rate=0.05)

# calc doublet rate, print sc.external.pl.scrublet_score_distribution to check.

sc.external.pl.scrublet_score_distribution(adata)

adata

adata.obs.predicted_doublet.unique().tolist()

#make a objet that still contains the predicted doublets for plotting later
ad=adata

l=[False]
adata=adata[adata.obs.predicted_doublet.isin(l)]

adata.write_h5ad(results_file)
